+++ CHANGELOG.md
@@ -6,10 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 [Semantic Versioning](http://semver.org/).
 
 ## 1.8.2.dev (development stage/unreleased)
+### Added 
+- Endpoints for www.binance.com margin UserData listenkey [issue #35](https://github.com/oliver-zehentleitner/unicorn-binance-websocket-api/issues/35)
+### Changed
+- Endpoints for www.binance.com spot UserData listenkey [issue #35](https://github.com/oliver-zehentleitner/unicorn-binance-websocket-api/issues/35)
 
 ## 1.8.2
 ### Fixed
-- Errors when creating private DEX streams: https://github.com/oliver-zehentleitner/unicorn-binance-websocket-api/issues/34
+- Errors when creating private DEX streams: [issue #34](https://github.com/oliver-zehentleitner/unicorn-binance-websocket-api/issues/34)
 
 ## 1.8.1
 ### Changed